requirements.txt
asgiref==3.4.1
awscli==1.20.15
boto3==1.18.14
botocore==1.21.15
cffi==1.14.6
colorama==0.4.3
cryptography==3.4.7
dj-database-url==0.5.0
Django==3.2.5
django-cors-headers==3.7.0
django-heroku==0.3.1
django-rest-knox==4.1.0
django-storages==1.11.1
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.7.2
docutils==0.15.2
gunicorn==20.1.0
jmespath==0.10.0
Pillow==8.3.1
psycopg2==2.9.1
pyasn1==0.4.8
pycparser==2.20
PyJWT==2.1.0
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
PyYAML==5.4.1
rsa==4.7.2
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
whitenoise==5.3.0
